fix cuda compilation errors, disable some tests
