+++ src/serac/physics/utilities/functional/boundary_integral_kernels.cuh
@@ -1,0 +1,8 @@
+// Copyright (c) 2019-2021, Lawrence Livermore National Security, LLC and
+// other Serac Project Developers. See the top-level LICENSE file for
+// details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+#pragma once
+
+// TODO: implement boundary integral kernels for CUDA device
